+++ src/execution/UserInterfaceTools.ts
@@ -15,11 +15,20 @@ export interface UserInterfaceTools {
 
 export interface UserInterfaceToolsPromptDialogOptions {
     /**
+     * Prompt title
+     *
+     * Note: This is not a prompt to language model but a prompt to the user
+     * @example "Your name"
+     */
+    promptTitle: string;
+
+    /**
      * Prompt message
      *
      * Note: This is not a prompt to language model but a prompt to the user
+     * @example "Please enter your name, including your last name, title, etc."
      */
-    prompt: string;
+    promptMessage: string;
 
     /**
      * Default value for the input/textarea
@@ -30,4 +39,11 @@ export interface UserInterfaceToolsPromptDialogOptions {
      * Placeholder for the input/textarea
      */
     placeholder?: string;
+
+    /**
+     * Priority of the prompt
+     *
+     * Note: This would be reflected for example into the UI z-index of the prompt modal
+     */
+    priority: string;
 }